requirements.txt
# Core PDF Processing
camelot-py[cv]>=0.11.0    # Table extraction - REQUIRES GHOSTSCRIPT
pdfplumber>=0.10.0         # Fallback table extraction
pymupdf>=1.23.0            # PDF handling, metadata extraction

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0            # Excel export

# Data Validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# LLM Integration
openai>=1.0.0              # GPT-4o
anthropic>=0.18.0          # Claude

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.0

# CLI and UI
click>=8.1.0               # CLI framework
rich>=13.0.0               # Beautiful terminal output
tqdm>=4.66.0               # Progress bars

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Development
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0
ruff>=0.1.0

# Optional: For API server
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic[email]>=2.0.0

# Optional: For monitoring
prometheus-client>=0.19.0